+++ src/shared/AkkaDotNet.Messages/Events/ItemActorEvents.cs
@@ -1,0 +1,7 @@
+﻿namespace AkkaDotNet.Messages.Events;
+
+public interface  IItemEvent : IWithItem{}
+
+public sealed record ItemAdded(string ItemId, int Count) : IItemEvent;
+
+public sealed record ItemRemoved(string ItemId, int Count) : IItemEvent;